+++ investments/client/Cargo.toml
@@ -5,7 +5,7 @@ edition = "2021"
 
 [dependencies]
 investments-program = { path = "../programs/investments", features = ["no-entrypoint"], version = "0.1.0" }
-serum-crank = { path = "../../serum_crank/programs/serum_crank", features = ["no-entrypoint"], version = "0.1.0" }
+openbook-crank = { path = "../../openbook_crank/programs/openbook_crank", features = ["no-entrypoint"], version = "0.1.0" }
 clockwork-client = "1.4.0"
 anchor-spl = { version = "0.26.0", features = ["dex"] }
 anchor-lang = "0.26.0"
