+++ openbook_crank/.gitignore
