+++ openbook_crank/Cargo.toml
@@ -1,0 +1,14 @@
+[workspace]
+members = [
+    "client",
+    "programs/*"
+]
+
+[profile.release]
+overflow-checks = true
+lto = "fat"
+codegen-units = 1
+[profile.release.build-override]
+opt-level = 3
+incremental = false
+codegen-units = 1
