+++ openbook_crank/client/Cargo.toml
@@ -1,10 +1,10 @@
 [package]
-name = "serum-crank-client"
+name = "openbook-crank-client"
 version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-serum-crank = { path = "../programs/serum_crank", features = ["no-entrypoint"], version = "0.1.0" }
+openbook-crank = { path = "../programs/openbook_crank", features = ["no-entrypoint"], version = "0.1.0" }
 clockwork-sdk = { version = "1.4.0" }
 clockwork-client = { version = "1.4.0" }
 anchor-spl = { version = "0.26.0", features = ["dex"] }
@@ -14,4 +14,4 @@ solana-cli-config = "1.10.4"
 solana-client = "1.13.5"
 
 [features]
-localnet = ["serum-crank/localnet"]
+localnet = ["openbook-crank/localnet"]
