+++ openbook_crank/package.json
@@ -4,12 +4,12 @@
         "lint": "prettier */*.js \"*/**/*{.js,.ts}\" --check"
     },
     "dependencies": {
-        "@project-serum/anchor": "^0.24.2"
+        "@project-serum/anchor": "^0.25.0"
     },
     "devDependencies": {
         "chai": "^4.3.4",
         "mocha": "^9.0.3",
-        "ts-mocha": "^8.0.0",
+        "ts-mocha": "^10.0.0",
         "@types/bn.js": "^5.1.0",
         "@types/chai": "^4.3.0",
         "@types/mocha": "^9.0.0",
