+++ openbook_crank/programs/openbook_crank/Cargo.toml
@@ -1,12 +1,12 @@
 [package]
-name = "serum-crank"
+name = "openbook-crank"
 version = "0.1.0"
 description = "Created with Anchor"
 edition = "2021"
 
 [lib]
 crate-type = ["cdylib", "lib"]
-name = "serum_crank"
+name = "openbook_crank"
 
 [features]
 no-entrypoint = []
