+++ openbook_crank/programs/openbook_crank/src/state/openbook_dex.rs
@@ -5,13 +5,6 @@ pub struct OpenBookDex;
 
 impl Id for OpenBookDex {
     fn id() -> Pubkey {
-        #[cfg(feature = "localnet")]
-        use std::str::FromStr;
-        #[cfg(feature = "localnet")]
-        // the program id of the openbook dex program as found in dex/serum_dex-keypair.json
-        return Pubkey::from_str("6QUbdf53eJZToaiLBrWbsJCE8jhXYAvufuJ4rzRskiCJ").unwrap();
-
-        #[cfg(not(feature = "localnet"))]
         anchor_spl::dex::ID
 
         // ORIGINAL OPENBOOK DEX CODE BELOW:
