--- serum_crank/Cargo.toml
@@ -1,5 +1,0 @@
-[workspace]
-members = [
-    "client",
-    "programs/*"
-]
