--- serum_crank/programs/serum_crank/src/id.rs
@@ -1,3 +1,0 @@
-use anchor_lang::prelude::declare_id;
-
-declare_id!("4rHBrNiBJBJUxeKd6bnmT31gGbRwW4baBZu3CXqE2h9q");
